Adressed easy to change review comments.
