+++ pkg/webhook/mutation/pod/attributes/attributes.go
@@ -129,3 +129,16 @@ func (attrs *PodAttributes) combineForJSONAnnotation() map[string]string {
 
 	return combined
 }
+
+type convertFunc func(string, string) string
+
+func convert(attributes map[string]string, c convertFunc) []string {
+	converted := make([]string, 0, len(attributes))
+	for key, value := range attributes {
+		if result := c(key, value); result != "" {
+			converted = append(converted, result)
+		}
+	}
+
+	return converted
+}
