+++ pkg/webhook/mutation/pod/attributes/podannotations_test.go
@@ -11,27 +11,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func TestSetPodAnnotationIfNotExists(t *testing.T) {
-	t.Run("sets key when absent", func(t *testing.T) {
-		pod := &corev1.Pod{ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{}}}
-		setPodAnnotationIfNotExists(pod, "my-key", "my-value")
-		assert.Equal(t, "my-value", pod.Annotations["my-key"])
-	})
-
-	t.Run("does not overwrite existing key", func(t *testing.T) {
-		pod := &corev1.Pod{ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{"my-key": "existing"}}}
-		setPodAnnotationIfNotExists(pod, "my-key", "new-value")
-		assert.Equal(t, "existing", pod.Annotations["my-key"])
-	})
-
-	t.Run("initializes nil annotations map before setting", func(t *testing.T) {
-		pod := &corev1.Pod{}
-		setPodAnnotationIfNotExists(pod, "my-key", "my-value")
-		require.NotNil(t, pod.Annotations)
-		assert.Equal(t, "my-value", pod.Annotations["my-key"])
-	})
-}
-
 func TestSetPodMetadataJSONAnnotation(t *testing.T) {
 	parseJSON := func(t *testing.T, pod *corev1.Pod) map[string]string {
 		t.Helper()
